#151. Moved BashContainer and CopyToClipboard components out of the sections folder.
